requirements.txt
amqp==5.2.0
asgiref==3.8.1
Baker==1.3
billiard==4.2.0
black==24.8.0
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
identify==2.6.0
iniconfig==2.0.0
kombu==5.4.0
model-bakery==1.19.4
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
